Update GitHub Actions workflow for version releases
The commit replaces the existing model for version releases with a more automated system that uses the CupOfTea696/gh-action-auto-release package for the action. This allows for automatic version bumping and tag creation using semver on every push to the main branch. The older process for creating GitHub Release and uploading executables has been removed.